requirements.txt
anthropic-claude-memory-system==1.0.0
chromadb>=0.4.22
openai>=1.0.0
sentence-transformers>=2.2.2
pyyaml>=6.0
python-dotenv>=1.0.0
tiktoken>=0.5.2
markdownify>=0.11.6
langchain>=0.0.340
langchain-community>=0.0.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
watchdog>=3.0.0
